simplify by using generator
