+++ concertina_helper/tune_on_layout.py
@@ -1,12 +1,12 @@
 from __future__ import annotations
 from dataclasses import dataclass
-from itertools import chain
 from collections.abc import Iterable
 
-from .layouts.bisonoric import BisonoricLayout, AnnotatedBisonoricFingering, Annotation
+from .layouts.bisonoric import BisonoricLayout, AnnotatedBisonoricFingering
 from .finger_finder import find_best_fingerings
 from .penalties import PenaltyFunction
-from .type_defs import Pitch
+from .type_defs import Annotation, Pitch
+from .note_generators import notes_from_tune
 
 from pyabc2 import Tune
 
@@ -24,24 +24,18 @@ class TuneOnLayout:
         '''
         For each note in the tune, returns all possible fingerings.
         '''
-        return list(chain(*[
-            [
-                (
-                    annotation := Annotation(
-                        measure=i + 1,
-                        pitch=(pitch := Pitch(note.to_pitch().name))
-                    ),
-                    {
-                        AnnotatedBisonoricFingering(
-                            fingering=f,
-                            annotation=annotation)
-                        for f in self.layout.get_fingerings(pitch)
-                    }
-                )
-                for note in measure
-            ]
-            for i, measure in enumerate(self.tune.measures)
-        ]))
+        return [
+            (
+                annotation,
+                {
+                    AnnotatedBisonoricFingering(
+                        fingering=f,
+                        annotation=annotation)
+                    for f in self.layout.get_fingerings(annotation.pitch)
+                }
+            )
+            for annotation in notes_from_tune(self.tune)
+        ]
 
     def get_best_fingerings(self, penalty_functions: Iterable[PenaltyFunction]) \
             -> Iterable[AnnotatedBisonoricFingering]:
@@ -49,9 +43,8 @@ class TuneOnLayout:
         Returns a list of fingerings that minimizes the cost for the entire tune,
         as measured by the provided `penalty_functions`.
         '''
-        all_fingerings = self.get_all_fingerings()
         f_sets = []
-        for annotation, f_set in all_fingerings:
+        for annotation, f_set in self.get_all_fingerings():
             if not f_set:
                 a = annotation
                 raise ValueError(f'No fingerings for {a.pitch} in measure {a.measure}')
